refactor: add canvas renderer state
