+++ packages/canvas/src/components/container/CanvasContainer.vue
@@ -119,6 +119,7 @@ export default {
         iframeMonitoring()
 
         initCanvas({ emit, renderer: detail, iframe: iframe.value, controller: props.controller })
+        useCanvas().renderer = detail
 
         const doc = iframe.value.contentDocument
         const win = iframe.value.contentWindow
