+++ packages/canvas/src/components/render/RenderMain.js
@@ -18,6 +18,8 @@ import { generateFunction } from '@opentiny/tiny-engine-controller/utils'
 import renderer, { parseData, setConfigure, setController, globalNotify, isStateAccessor } from './render'
 import { getNode as getNodeById, clearNodes, getRoot, setContext, getContext, setCondition, context } from './context'
 import CanvasEmpty from './CanvasEmpty.vue'
+import { getCurrent, setLocales, updateRect, addStyle, addScript, canvasDispatch } from '../container/container'
+import Builtin from '../builtin/builtin.json'
 
 const { BROADCAST_CHANNEL } = constants
 
@@ -422,7 +424,15 @@ export const api = {
   getGlobalState,
   getDataSourceMap,
   setDataSourceMap,
-  setGlobalState
+  setGlobalState,
+  getCurrent, 
+  setLocales,
+  getNodeById: getNode,
+  updateRect,
+  addStyle, 
+  addScript, 
+  canvasDispatch
 }
 
 window.api = api
+window.Builtin = Builtin
