+++ packages/controller/src/useCanvas.js
@@ -11,8 +11,7 @@
  */
 
 /* eslint-disable no-new-func */
-import { reactive } from 'vue'
-import { setSchema, renderApi } from '@opentiny/tiny-engine-canvas'
+import { reactive, ref } from 'vue'
 import { constants } from '@opentiny/tiny-engine-utils'
 import useHistory from './useHistory'
 
@@ -52,12 +51,14 @@ const defaultSchema = {
   outputs: []
 }
 
+const renderer = ref(null)
+
 const pageState = reactive({ ...defaultPageState, loading: true })
 // 重置画布数据
 const resetCanvasState = async (state = {}) => {
   Object.assign(pageState, defaultPageState, state)
 
-  await setSchema(pageState.pageSchema)
+  await renderer.value?.setSchema(pageState.pageSchema)
 }
 
 // 页面重置画布数据
@@ -141,6 +142,7 @@ const getCurrentPage = () => pageState.currentPage
 export default function () {
   return {
     pageState,
+    renderer,
     isBlock,
     isSaved,
     isLoading,
@@ -151,8 +153,8 @@ export default function () {
     resetPageCanvasState,
     resetBlockCanvasState,
     clearCurrentState,
-    getDataSourceMap: renderApi.getDataSourceMap,
-    setDataSourceMap: renderApi.setDataSourceMap,
+    getDataSourceMap: renderer.value?.getDataSourceMap,
+    setDataSourceMap: renderer.value?.setDataSourceMap,
     getCurrentSchema,
     setCurrentSchema,
     getCurrentPage
