+++ packages/settings/design/package.json
@@ -25,6 +25,7 @@
   "homepage": "https://opentiny.design/tiny-engine",
   "dependencies": {
     "@opentiny/tiny-engine-common": "workspace:*",
+    "@opentiny/tiny-engine-controller": "workspace:*",
     "@opentiny/vue": "~3.10.0",
     "sortablejs": "^1.14.0"
   },
